Refactor the testing setup.
Refactor the testing setup, dropping the globaly required protractor, and run the testing tasks with Grunt and a locally installed protractor.
The tests will now run on the production build.

Refactor the background.spec tests to get the extension ID dinamically, fixing the long standing issues with the hardcoded extension options page url.

Fix all failing tests in the the background and content specs, so all tests are now passing with `grunt test`.

Added details about how to run the new testing setup in the README.