Show more fields: Subject, To, CC, BCC (#219)
* Initial work on the show-more-fields functionality in options.

* Refactor the `setTitle` plugin method as `setField`. It will be called for each available field in the quicktext (subject, to, etc.).
Initial support for TO in the gmail plugin.

* Add support for to, cc and bcc in background.
Create a flex grid for the to, cc, bcc fields in the form. Visibile fields will take up the space equally.
Don't save fields with no actual content in them.

* Add support for cc and bcc autocomplete in the gmail plugin.

* setField is not mandatory, so don't add it to the blank default plugin.

* Add subject support to the outlook plugin.
Outlook support was broken because they changed the domain name from `mail.live.com` to `outlook.live.com`.

* Add outlook support for To, CC and BCC, only in New Message.

* Backwards compability for quicktexts saved with empty `` subjects.
Remove the empty subject from the default quicktext on `new`.
These caused the subject field to always be visible.

* Refactor the `setField` plugin method to a more generic `before` method.
The `before` method will be run right before inserting the quicktext, after the plugin's `getData`.
This was required for outlook support, and also seems like a better idea overall.
Also added support for an `after` method in plugins, that is ran after the quicktext is inserted - but it currently unused by any plugins.

* Remove `setTitle` from the facebook plugin.

* Remove `setTitle` from the `desk` plugin.

* Initial work on supporting to,cc,bcc in Fastmail.

* Add support for to/cc/bcc in the fastmail plugin.

* Adapt the linkedin plugin to use the new `before` method.

* Tweaks to the fastmail plugin to support message reply, and not hide already visible cc and bcc fields.

* Support for to/cc/bcc in the Yahoo plugin.

* Adapt the uservoice plugin to use the `before` method instead of `setTitle`.

* Remove `setTitle` from the zendesk plugins.

* Remove `setTitle` from the Yahoo plugin.

* Remove `setField` from the outlook plugin.